event_scraper_tool/requirements.txt
streamlit
openai
pandas
beautifulsoup4
selenium
python-dateutil
